+++ libs/location/feature/src/lib/location-list.store.ts
@@ -1,11 +1,14 @@
 import { computed, inject } from '@angular/core';
 import { rxResource } from '@angular/core/rxjs-interop';
-import { ModalController } from '@ionic/angular/standalone';
+import { ModalController, ToastController } from '@ionic/angular/standalone';
 import { patchState, signalStore, withComputed, withMethods, withProps, withState } from '@ngrx/signals';
 
 import { AppStore } from '@bk2/shared-feature';
 import { LocationModel } from '@bk2/shared-models';
 import { chipMatches, nameMatches } from '@bk2/shared-util-core';
+import { bkTranslate } from '@bk2/shared-i18n';
+import { copyToClipboard, showToast } from '@bk2/shared-util-angular';
+import { MapViewModalComponent } from '@bk2/shared-ui';
 
 import { LocationService } from '@bk2/location-data-access';
 import { isLocation } from '@bk2/location-util';
@@ -29,7 +32,8 @@ export const LocationListStore = signalStore(
   withProps(() => ({
     locationService: inject(LocationService),
     appStore: inject(AppStore),
-    modalController: inject(ModalController),    
+    modalController: inject(ModalController),
+    toastController: inject(ToastController),
   })),
   withProps((store) => ({
     locationsResource: rxResource({
@@ -80,60 +84,63 @@ export const LocationListStore = signalStore(
       },
 
       /******************************* actions *************************************** */
-      async add(): Promise<void> {
+      async add(readOnly = true): Promise<void> {
         const location = new LocationModel(store.appStore.tenantId());
+        await this.edit(location, readOnly);
+      },
+
+      async edit(location: LocationModel, readOnly = true): Promise<void> {
         const modal = await store.modalController.create({
           component: LocationEditModalComponent,
           componentProps: {
             location: location,
-            currentUser: store.currentUser()
+            currentUser: store.currentUser(),
+            readOnly
           }
         });
         modal.present();
         const { data, role } = await modal.onDidDismiss();
-        if (role === 'confirm') {
+        if (role === 'confirm' && data) {
           if (isLocation(data, store.appStore.tenantId())) {
-            await store.locationService.create(data, store.currentUser());
+            data.bkey?.length > 0 ?
+              await store.locationService.update(data, store.currentUser()) : 
+              await store.locationService.create(data, store.currentUser());
           }
         }
         store.locationsResource.reload();
       },
 
-      async delete(location: LocationModel): Promise<void> {
+      async delete(location: LocationModel, readOnly = true): Promise<void> {
+        if (readOnly) return;
         await store.locationService.delete(location, store.currentUser());
         this.reset();
-      },
-
-      async edit(location: LocationModel): Promise<void> {
-        const modal = await store.modalController.create({
-          component: LocationEditModalComponent,
-          componentProps: {
-            location: location,
-            currentUser: store.currentUser()
-          }
-        });
-        modal.present();
-        const { data, role } = await modal.onDidDismiss();
-        if (role === 'confirm') {
-          if (isLocation(data, store.appStore.tenantId())) {
-            await store.locationService.update(data, store.currentUser());
-          }
-        }
-        store.locationsResource.reload();
       },
 
       async export(type: string): Promise<void> {
         console.log(`LocationListStore.export(${type}) is not yet implemented.`);
       },
 
-      async show(location: LocationModel): Promise<void> {
+      async showOnMap(location: LocationModel): Promise<void> {
         console.log('LocationListStore.show is not yet implemented.');
+        const modal = await store.modalController.create({
+          component: MapViewModalComponent,
+          componentProps: {
+            title: location.name,
+            initialPosition: {
+              lat: location.latitude,
+              lng: location.longitude,
+            },
+            zoom: 15,
+            enableTrafficLayer: false
+          }
+        });
+        modal.present();
+        await modal.onWillDismiss();
       },
 
       async copy(location: LocationModel): Promise<void> {
-        console.log('LocationListStore.copy is not yet implemented.');
+        await copyToClipboard(location.latitude + ', ' + location.longitude);
+        await showToast(store.toastController, bkTranslate('@location.operation.copy.conf'));
       },
   }})
 );
-
-
